-attempt03

+++ meme/meme.c
@@ -178,7 +178,7 @@ static ssize_t meme_write(struct file* file, const char __user* buf, size_t SIZE
 
 static int meme_increment(struct meme_increment_t __user *arg)
 {
-	int meme = *arg;
+	uint64_t meme = *arg;
 	meme++;
 	
 	return meme;
